pyproject.toml
[project]
name = "evosite3d"
version = "0.1"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.85",
    "matplotlib",
    "pandas",
    "scipy",
    "seaborn",
    "statsmodels",
    "torch==2.5.1",
    "torchaudio==2.5.1",
    "torchdata==0.8.0",
    "torchvision==0.20.1",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=7.3.0",
    "mypy>=1.5.0",
    "pylint>=3.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.pylint.format]
max-line-length = "100"

[tool.setuptools.packages.find]
exclude = ["outputs*", "tutorials*", "ncbi_data*", "test_outputs*"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
